requirements.txt
docker == 3.3.0
PyYAML == 3.12
pathlib == 1.0.1